Network Security/requirements.txt
pandas
numpy
scikit-learn    
matplotlib
seaborn
python-dotenv
flask
pymongo
pymongo[srv]==3.12
certifi
pyyaml
mlflow
fastapi
uvicorn
dnspython
python-multipart
#-e . 
